fix: resolve bug de duplicação das divs
