+++ src/modules/content/helpers/checkNaN.ts
@@ -1,0 +1,3 @@
+export function checkNaN(hr: string) {
+  return hr.includes("NaN") ? "NOT FOUND" : hr;
+}
