--- src/modules/content/layout/totalClosed.ts
@@ -1,5 +1,0 @@
-export const updateTotalClosedWrapper = (totalClosed: number) => {
-  const totalClosedWrapper = document.querySelector(".summary-closed-tasks");
-  const totalClosedNumber = totalClosedWrapper.childNodes[0] as HTMLElement;
-  totalClosedNumber.innerText = `${totalClosed}`;
-};
